apps/api/functions/requirements.txt
firebase_functions~=0.1.0
livekit-api
flask
python-dotenv